fix: only validate modules directory if required (e.g. for install)
