+++ .changeset/itchy-horses-wink.md
@@ -1,0 +1,5 @@
+---
+"@pnpm/get-context": major
+---
+
+Don't validate (and possibly purge) modules directory as a side effect of `getContext` and `getContextForSingleImporter` [#8657](https://github.com/pnpm/pnpm/pull/8657).
