+++ .changeset/perfect-spoons-call.md
@@ -1,0 +1,5 @@
+---
+"@pnpm/get-context": major
+---
+
+`PnpmSingleContext.hoistPattern` and `PnpmSingleContext.publicHoistPattern` are no longer affected by modules directory state [#8657](https://github.com/pnpm/pnpm/pull/8657).
