+++ .changeset/spicy-apricots-beam.md
@@ -1,0 +1,5 @@
+---
+"@pnpm/get-context": major
+---
+
+`UnexpectedStoreError` and `UnexpectedVirtualStoreDirError` are no longer exported [#8657](https://github.com/pnpm/pnpm/pull/8657). They can be imported from `@pnpm/core` instead.
