+++ .changeset/tiny-terms-help.md
@@ -1,0 +1,5 @@
+---
+"@pnpm/get-context": major
+---
+
+Argument `alreadyPurged` removed from `getContextForSingleImporter` [#8657](https://github.com/pnpm/pnpm/pull/8657).
