+++ pkg-manager/core/src/index.ts
@@ -9,7 +9,9 @@ export type {
 export type { HoistingLimits } from '@pnpm/headless'
 export * from './api'
 
-export { type ProjectOptions, UnexpectedStoreError, UnexpectedVirtualStoreDirError } from '@pnpm/get-context'
+export { type ProjectOptions } from '@pnpm/get-context'
+export { UnexpectedStoreError } from './install/checkCompatibility/UnexpectedStoreError'
+export { UnexpectedVirtualStoreDirError } from './install/checkCompatibility/UnexpectedVirtualStoreDirError'
 export type { InstallOptions } from './install/extendInstallOptions'
 
 export type { WorkspacePackages } from '@pnpm/resolver-base'
