+++ pkg-manager/core/src/install/checkCompatibility/BreakingChangeError.ts
