+++ pkg-manager/core/src/install/checkCompatibility/ErrorRelatedSources.ts
