+++ pkg-manager/core/src/install/checkCompatibility/ModulesBreakingChangeError.ts
