+++ pkg-manager/core/src/install/checkCompatibility/UnexpectedStoreError.ts
