+++ pkg-manager/core/src/install/checkCompatibility/UnexpectedVirtualStoreDirError.ts
