+++ pkg-manager/core/src/install/checkCompatibility/index.ts
