+++ pkg-manager/get-context/package.json
@@ -40,15 +40,12 @@
   "dependencies": {
     "@pnpm/constants": "workspace:*",
     "@pnpm/core-loggers": "workspace:*",
-    "@pnpm/error": "workspace:*",
     "@pnpm/lockfile.fs": "workspace:*",
     "@pnpm/modules-yaml": "workspace:*",
     "@pnpm/read-projects-context": "workspace:*",
     "@pnpm/resolver-base": "workspace:*",
     "@pnpm/types": "workspace:*",
-    "@zkochan/rimraf": "catalog:",
     "ci-info": "catalog:",
-    "enquirer": "catalog:",
     "path-absolute": "catalog:",
     "ramda": "catalog:"
   },
